embeddings-service/requirements.txt
transformers>=4.37.0
sentencepiece
protobuf
torch
pillow
flask
uvicorn
fastapi
python-multipart